Ex_Files_Java_Build_Automation_Maven/Exercise Files/ch06/06_04/calculator/src/main/java/com/lynda/App.java
package com.lynda;


public class App {
	public static void main(String[] args){
		
		System.out.println(sum(10,5));
		System.out.println(subtraction(10,5));
		System.out.println(division(10,5));
		System.out.println(multiplication(10,5));

	}
	public static int sum(int a, int b)
	{
		return a + b;
	}
	public static int subtraction(int a, int b)
	{
		return a - b;
	}
	public static int division(int a, int b)
	{
		return a / b;
	}
	public static int multiplication(int a, int b)
	{
		return a * b;
	}
	
}
